move pkgs, update readme, add two more metrics to influx client
