+++ examples/up.xml.example
@@ -79,7 +79,7 @@
   <unifi_pass></unifi_pass>
   <unifi_url>https://127.0.0.1:8443</unifi_url>
   <!--
-  # Enable collection of Intrusion Detection System Data.
+  # Enable collection of Intrusion Detection System Data (InfluxDB only).
   # Only useful if IDS or IPS are enabled on one of the sites.
   -->
   <collect_ids>false</collect_ids>
