+++ examples/up.yaml.example
@@ -57,7 +57,7 @@ unifi_user: "influx"
 unifi_pass: ""
 unifi_url: "https://127.0.0.1:8443"
 
-# Enable collection of Intrusion Detection System Data.
+# Enable collection of Intrusion Detection System Data (InfluxDB only).
 # Only useful if IDS or IPS are enabled on one of the sites.
 collect_ids: false
