+++ pkg/influxunifi/clients.go
@@ -56,6 +56,7 @@ func ClientPoints(c *unifi.Client, now time.Time) ([]*influx.Point, error) {
 		"essid":                  c.Essid,
 		"bssid":                  c.Bssid,
 		"radio_desc":             c.RadioDescription,
+		"satisfaction":           c.Satisfaction.Val,
 		"hostname":               c.Hostname,
 		"dpi_stats_last_updated": c.DpiStatsLastUpdated,
 		"last_seen_by_uap":       c.LastSeenByUAP,
@@ -84,6 +85,7 @@ func ClientPoints(c *unifi.Client, now time.Time) ([]*influx.Point, error) {
 		"tx_bytes":               c.TxBytes,
 		"tx_bytes_r":             c.TxBytesR,
 		"tx_packets":             c.TxPackets,
+		"tx_retries":             c.TxRetries,
 		"tx_power":               c.TxPower,
 		"tx_rate":                c.TxRate,
 		"uptime":                 c.Uptime,
